Split screen
I have made split screen with keyboard and gamepad controls.
please note, apparently local multiplayer is very dificult to accomplish
in ue4 in the sense of controllers, the hope right now that is untested
due to only having 1 controller is that if a second controller is
plugged in the second player will move accordingly please test this and
get back to me asap so i can plan around it.

also important, if you play from the viewport it opens up a client
window you have to ult tab back to the view port. however if you play
from standalone it opens to a separate window but no editor options.
thats all